Move LandingPage to pages/ and add auth checking to redirect to inbox
